#14-Generating_Game_Board-moved graph in class
- Created a new Class to handle the graph Creation
- Clears up Game Board Class
